+++ src/test/resources/ru/vyarus/gradle/plugin/animalsniffer/java/invalid/Sample2.java
@@ -1,5 +1,7 @@
 package invalid;
 
+import org.junit.Assert;
+
 import java.nio.file.FileSystems;
 
 public class Sample2 {
@@ -7,5 +9,6 @@ public class Sample2 {
     public static void main(String[] args) {
         // added in 1.7
         FileSystems.getDefault().getFileStores();
+        Assert.assertTrue(true);
     }
 }